Fix e2e code triggering deadcode error on mac
Split MacOS / Linux e2e Privileged func
Address e2e funcs that use undefined MS_BIND on non-Linux
Add missing e2e/internal/e2e/home.go return
