+++ e2e/internal/e2e/priv_linux.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019, Sylabs Inc. All rights reserved.
+// Copyright (c) 2019,2020 Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -6,19 +6,11 @@
 package e2e
 
 import (
-	"os"
 	"runtime"
 	"syscall"
 	"testing"
 
 	"github.com/pkg/errors"
-)
-
-var (
-	// uid of original user running test.
-	origUID = os.Getuid()
-	// gid of original group running test.
-	origGID = os.Getgid()
 )
 
 // Privileged wraps the supplied test function with calls to ensure
@@ -52,14 +44,4 @@ func Privileged(f func(*testing.T)) func(*testing.T) {
 
 		f(t)
 	}
-}
-
-// OrigUID returns the UID of the user running the test suite.
-func OrigUID() int {
-	return origUID
-}
-
-// OrigGID returns the GID of the user running the test suite.
-func OrigGID() int {
-	return origGID
 }
